Catboost-sagemaker/requirements.txt
eventlet
greenlet
pandas
catboost
pyyaml
mlflow
numpy==1.16.2 
scipy==1.2.1 
scikit-learn==0.20.2 
pandas 
flask 
gevent 
gunicorn
seaborn